Module pallet_moonbeam_orbiters::pallet::dispatchables

Auto-generated docs-only module listing all defined dispatchables for this pallet.

Warning: Doc-Only

Members of this module cannot be used directly and are only provided for documentation purposes. To see the real version of each dispatchable, look for them in Pallet or Call.

Functions

add_collator

Add a collator to orbiters program.

collator_add_orbiter

Add an orbiter in a collator pool

collator_remove_orbiter

Remove an orbiter from the caller collator pool

orbiter_leave_collator_pool

Remove the caller from the specified collator pool

orbiter_register

Registering as an orbiter

orbiter_unregister

Deregistering from orbiters

remove_collator

Remove a collator from orbiters program.