Module pallet_moonbeam_orbiters::pallet::storage_types

Auto-generated docs-only module listing all defined storage types for this pallet. Note that members of this module cannot be used directly and are only provided for documentation purposes.

Structs

AccountLookupOverride

Account lookup override

CollatorsPool

Current orbiters, with their “parent” collator

CurrentRound

Current round index

ForceRotation

If true, it forces the rotation at the next round. A use case: when changing RotatePeriod, you need a migration code that sets this value to true to avoid holes in OrbiterPerRound.

MinOrbiterDeposit

Minimum deposit required to be registered as an orbiter

OrbiterPerRound

Store active orbiter per round and per parent collator

RegisteredOrbiter

Check if account is an orbiter