Crate pallet_moonbeam_orbiters

Pallet moonbeam-orbiters

This pallet allows authorized collators to share their block creation rights and rewards with multiple entities named “orbiters”. Each authorized collator will define a group of orbiters, and each orbiter will replace the collator in turn with the other orbiters (rotation every RotatePeriod rounds).

This pallet is designed to work with the nimbus consensus. In order not to impact the other pallets (notably nimbus and parachain-staking) this pallet simply redefines the lookup NimbusId-> AccountId, in order to replace the collator by its currently selected orbiter.

Re-exports

pub use weights::WeightInfo;

pub use pallet::*;

pub use types::*;

Modules

pallet

The pallet module in each FRAME pallet hosts the most important items needed to construct this pallet.

types

weights

Autogenerated weights for pallet_moonbeam_orbiters