Trait pallet_moonbeam_orbiters::pallet::Config

pub trait Config: Config {
    type Event: From<Event<Self>> + IsType<<Self as Config>::Event>;
    type AccountLookup: AccountLookup<Self::AccountId>;
    type AddCollatorOrigin: EnsureOrigin<Self::Origin>;
    type Currency: NamedReservableCurrency<Self::AccountId>;
    type DelCollatorOrigin: EnsureOrigin<Self::Origin>;
    type MaxPoolSize: Get<u32>;
    type MaxRoundArchive: Get<Self::RoundIndex>;
    type OrbiterReserveIdentifier: Get<ReserveIdentifierOf<Self>>;
    type RotatePeriod: Get<Self::RoundIndex>;
    type RoundIndex: Parameter + Member + MaybeSerializeDeserialize + Debug + Default + MaybeDisplay + AtLeast32Bit + Copy;
    type WeightInfo: WeightInfo;
}

Configuration trait of this pallet.

Implement this type for a runtime in order to customize this pallet.

Associated Types

type Event: From<Event<Self>> + IsType<<Self as Config>::Event>

Overarching event type.

type AccountLookup: AccountLookup<Self::AccountId>

A type to convert between AuthorId and AccountId. This pallet wrap the lookup to allow orbiters authoring.

type AddCollatorOrigin: EnsureOrigin<Self::Origin>

Origin that is allowed to add a collator in orbiters program.

type Currency: NamedReservableCurrency<Self::AccountId>

The currency type.

type DelCollatorOrigin: EnsureOrigin<Self::Origin>

Origin that is allowed to remove a collator from orbiters program.

type MaxPoolSize: Get<u32>

Maximum number of orbiters per collator.

type MaxRoundArchive: Get<Self::RoundIndex>

Maximum number of round to keep on storage.

type OrbiterReserveIdentifier: Get<ReserveIdentifierOf<Self>>

Reserve identifier for this pallet instance.

type RotatePeriod: Get<Self::RoundIndex>

Number of rounds before changing the selected orbiter.

type RoundIndex: Parameter + Member + MaybeSerializeDeserialize + Debug + Default + MaybeDisplay + AtLeast32Bit + Copy

Round index type.

type WeightInfo: WeightInfo

Weight information for extrinsics in this pallet.