Trait pallet_moonbeam_orbiters::pallet::Config

pub trait Config: Config {
    type RuntimeEvent: From<Event<Self>> + IsType<<Self as Config>::RuntimeEvent>;
    type AccountLookup: AccountLookup<Self::AccountId>;
    type AddCollatorOrigin: EnsureOrigin<Self::RuntimeOrigin>;
    type Currency: NamedReservableCurrency<Self::AccountId>;
    type DelCollatorOrigin: EnsureOrigin<Self::RuntimeOrigin>;
    type MaxPoolSize: Get<u32>;
    type MaxRoundArchive: Get<Self::RoundIndex>;
    type OrbiterReserveIdentifier: Get<ReserveIdentifierOf<Self>>;
    type RotatePeriod: Get<Self::RoundIndex>;
    type RoundIndex: Parameter + Member + MaybeSerializeDeserialize + Debug + Default + MaybeDisplay + AtLeast32Bit + Copy;
    type WeightInfo: WeightInfo;
}

Configuration trait of this pallet.

The main purpose of this trait is to act as an interface between this pallet and the runtime in which it is embedded in. A type, function, or constant in this trait is essentially left to be configured by the runtime that includes this pallet.

Consequently, a runtime that wants to include this pallet must implement this trait.

Required Associated Types

type RuntimeEvent: From<Event<Self>> + IsType<<Self as Config>::RuntimeEvent>

Overarching event type.

type AccountLookup: AccountLookup<Self::AccountId>

A type to convert between AuthorId and AccountId. This pallet wrap the lookup to allow orbiters authoring.

type AddCollatorOrigin: EnsureOrigin<Self::RuntimeOrigin>

Origin that is allowed to add a collator in orbiters program.

type Currency: NamedReservableCurrency<Self::AccountId>

The currency type.

type DelCollatorOrigin: EnsureOrigin<Self::RuntimeOrigin>

Origin that is allowed to remove a collator from orbiters program.

type MaxPoolSize: Get<u32>

Maximum number of orbiters per collator.

type MaxRoundArchive: Get<Self::RoundIndex>

Maximum number of round to keep on storage.

type OrbiterReserveIdentifier: Get<ReserveIdentifierOf<Self>>

Reserve identifier for this pallet instance.

type RotatePeriod: Get<Self::RoundIndex>

Number of rounds before changing the selected orbiter. WARNING: when changing RotatePeriod, you need a migration code that sets ForceRotation to true to avoid holes in OrbiterPerRound.

type RoundIndex: Parameter + Member + MaybeSerializeDeserialize + Debug + Default + MaybeDisplay + AtLeast32Bit + Copy

Round index type.

type WeightInfo: WeightInfo

Weight information for extrinsics in this pallet.

Object Safety

This trait is not object safe.